reorganise and remove old files
