--- CloudyLauncher.fxml
@@ -1,149 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.net.URL?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.Tab?>
-<?import javafx.scene.control.TabPane?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.TilePane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.text.Text?>
-<?import javafx.scene.layout.Region?>
-
-<VBox
-    fx:controller="CloudyLauncher"
-    xmlns:fx="http://javafx.com/fxml"
-    fx:id="rootLayout"
-    styleClass="root">
-    <children>
-        <TabPane fx:id="manageUserPanel">
-            <tabs>
-                <Tab
-                    text="Sign Up"
-                    closable="false">
-                    <VBox styleClass="user-panel">
-                        <GridPane styleClass="grid-pane">
-                            <Label
-                                text="Email"
-                                GridPane.columnIndex="0"
-                                GridPane.rowIndex="0" />
-                            <TextField
-                                fx:id="signupEmail"
-                                promptText="Enter email"
-                                GridPane.columnIndex="1"
-                                GridPane.rowIndex="0"
-                                onAction="#handleSignUp" />
-                            <Label
-                                text="Username"
-                                GridPane.columnIndex="0"
-                                GridPane.rowIndex="1" />
-                            <TextField
-                                fx:id="signupUsername"
-                                promptText="Enter username"
-                                GridPane.columnIndex="1"
-                                GridPane.rowIndex="1"
-                                onAction="#handleSignUp"/>
-                            <Label
-                                text="Password"
-                                GridPane.columnIndex="0"
-                                GridPane.rowIndex="2" />
-                            <PasswordField
-                                fx:id="signupPassword"
-                                promptText="Enter password"
-                                GridPane.columnIndex="1"
-                                GridPane.rowIndex="2"
-                                onAction="#handleSignUp" />
-                        </GridPane>
-
-                        <Button
-                            text="Sign Up"
-                            onAction="#handleSignUp" />
-                        <Text fx:id="signupFeedback" styleClass="feedback" wrappingWidth="290" />
-
-                    </VBox>
-                </Tab>
-                <Tab
-                    text="Login"
-                    closable="false">
-                    <VBox styleClass="user-panel">
-                        <GridPane styleClass="grid-pane">
-                            <Label
-                                text="Username"
-                                GridPane.columnIndex="0"
-                                GridPane.rowIndex="1" />
-                            <TextField
-                                fx:id="loginUsername"
-                                promptText="Enter username"
-                                GridPane.columnIndex="1"
-                                GridPane.rowIndex="1"
-                                onAction="#handleLogin" />
-                            <Label
-                                text="Password"
-                                GridPane.columnIndex="0"
-                                GridPane.rowIndex="2" />
-                            <PasswordField
-                                fx:id="loginPassword"
-                                promptText="Enter password"
-                                GridPane.columnIndex="1"
-                                GridPane.rowIndex="2"
-                                onAction="#handleLogin"/>
-                        </GridPane>
-                        <Button
-                            text="Login"
-                            onAction="#handleLogin" />
-                        <Text fx:id="loginFeedback" styleClass="feedback"  wrappingWidth="290" />
-                    </VBox>
-                </Tab>
-            </tabs>
-        </TabPane>
-        <HBox
-            fx:id="gameDisplayLayout"
-            styleClass="game-display-layout"
-            onMouseClicked="#handleDisplayGameInfo">
-            <VBox fx:id="tilePaneBase" styleClass="game-tile-layout">
-                <children>
-                    <HBox styleClass="buttons-display">
-                        <Button
-                            text="All Games"
-                            onAction="#handleDisplayAllGames" />
-                        <Button
-                            text="My Games"
-                            onAction="#handleDisplayMyGames" />
-                        <Button
-                            styleClass="refresh-button"
-                            onAction="#handleRefreshGameList" />
-                        <Region HBox.hgrow="ALWAYS"/>
-                        <Button
-                            text="Log Out"
-                            onAction="#handleLogOut" />
-                    </HBox>
-                    <TilePane fx:id="gameRoot" styleClass="game-root">
-                    </TilePane>
-
-                </children>
-            </VBox>
-            <VBox fx:id="gameInfoPanel" styleClass="game-info">
-
-                <Text fx:id="gameName" styleClass="game-info-name"/>
-                <Button fx:id="gameButton"
-                    text="Join Game"
-                    onAction="#handleGame" />
-                <Text fx:id="joinFeedback" styleClass="feedback"/>
-                <Text fx:id="gameInfo" styleClass="feedback"/>
-            </VBox>
-            <stylesheets>
-                <URL value="@style.css" />
-            </stylesheets>
-
-        </HBox>
-    </children>
-    <stylesheets>
-        <URL value="@style.css" />
-    </stylesheets>
-</VBox>
