+++ CloudyLauncher.java
@@ -1,241 +1,363 @@
-
-import java.io.BufferedReader;
-import java.io.DataOutputStream;
+import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
+import java.util.Map;
+import java.util.Properties;
 
 import javafx.application.Application;
-import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
+import javafx.scene.Node;
 import javafx.scene.control.Button;
-import javafx.scene.control.TabPane;
+import javafx.scene.control.Pagination;
 import javafx.scene.control.TextField;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.MouseEvent;
-import javafx.scene.layout.HBox;
-import javafx.scene.layout.Region;
+import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.StackPane;
 import javafx.scene.layout.TilePane;
 import javafx.scene.layout.VBox;
 import javafx.scene.text.Text;
-import javafx.scene.Node;
-import javafx.scene.Scene;
 import javafx.stage.Stage;
+import javafx.util.Callback;
 
 public class CloudyLauncher extends Application {
 
-    @FXML private VBox rootLayout;
-    @FXML private TabPane manageUserPanel;
-    @FXML private HBox gameDisplayLayout;
+    private CloudyLauncherServerInterface server;
+    private CloudyLauncherJsonParser parser = CloudyLauncherJsonParser.getParser();
 
-    @FXML private TextField signupEmail;
-    @FXML private TextField signupUsername;
-    @FXML private Text signupFeedback;
-
-    @FXML private TextField signupPassword;
-    @FXML private TextField loginUsername;
-    @FXML private TextField loginPassword;
-    @FXML private Text loginFeedback;
-
-    @FXML private TilePane gameRoot;
-    @FXML private VBox gameInfoPanel;
-    @FXML private VBox tilePaneBase;
-    @FXML private Text gameName;
-    @FXML private Text gameInfo;
-    @FXML private Button gameButton;
-    @FXML private Text joinFeedback;
-
-    private double infoWidth = 250;
-
-    private String baseurl = "http://127.0.0.1:8000";
-    private String token = "";
-    private String feedback = "";
+    private String token;
     private List<Game> listOfGames = new ArrayList<Game>();
-    private List<String> listOfOwnedIds = new ArrayList<String>();
+    private List<Game> listOfOwnedGames = new ArrayList<Game>();
     private Game selectedGame;
-    static Stage userStage = new Stage();
-    static Stage gameStage = new Stage();
 
-    @FXML
-    protected void handleSignUp(ActionEvent event) {
-        attemptUserRegistration(signupUsername.getText(),
-                                signupPassword.getText(),
-                                signupEmail.getText());
-        signupFeedback.setText(feedback);
+    @FXML private BorderPane mainContent;
+    @FXML private Pagination pagination;
+    @FXML private VBox gameInfoPanel;
+
+    @FXML private Button mainButton;
+    @FXML private TextField email;
+    @FXML private TextField username;
+    @FXML private TextField password;
+    @FXML private Text accountsFeedback;
+
+    @FXML private Text gameTitle;
+    @FXML private ImageView gameImage;
+    @FXML private Text gameTextInformation;
+    @FXML private Button gameButton;
+    @FXML private Text gameFeedback;
+
+    final private String PATH_LAUNCHER_BASE = "design/CL.fxml";
+    final private String PATH_ACCOUNTS_LOGIN = "design/Login.fxml";
+    final private String PATH_ACCOUNTS_SIGNUP = "design/Signup.fxml";
+    final private String PATH_GAME_DISPLAY_BASE = "design/GameDisplay.fxml";
+    final private String PATH_GAME_DISPLAY_INITIAL = "design/GameDisplay_initial.fxml";
+    final private String PATH_GAME_DISPLAY_GAME = "design/GameDisplay_game.fxml";
+
+    final private String ERROR_FXML_CONFIG = "ERROR: Cannot load settings/config.xml";
+    final private String ERROR_FXML_SIGNUP = "ERROR: Cannot load sign up page";
+    final private String ERROR_FXML_LOGIN = "ERROR: Cannot load login page";
+    final private String ERROR_FXML_GAME_DISPLAY = "ERROR: Cannot load game display";
+    final private String ERROR_GAME_JOIN = "ERROR: Cannot join game";
+
+    final private String URL_OWNED_BADGE = "images/orangeribbon.png";
+    final private String GAME_INFORMATION = "%s\n%s\n%s";
+
+    final private String COMMAND_RUN_THINCLIENT = "python thin_client/main.py %s %s %s %s";
+
+    final private Integer GAME_DISPLAY_WELCOME = 0;
+    final private Integer GAME_DISPLAY_GAME_INFO = 1;
+    final private Integer PAGINATION_ALL_GAMES = 2;
+    final private Integer PAGINATION_MY_GAMES = 3;
+    final private Integer TILES_PER_PAGE = 12;
+
+    public void setupConfig() {
+
+        try {
+            FileInputStream configFile = new FileInputStream("settings/config.xml");
+            Properties props = new Properties();
+            props.loadFromXML(configFile);
+            configFile.close();
+
+            String url = "http://%s:%s/";
+            String serverIp = props.getProperty("serverIp");
+            String serverPort = props.getProperty("serverPort");
+            server = new CloudyLauncherServerInterface(String.format(url,
+                                                                     serverIp,
+                                                                     serverPort));
+
+        } catch (IOException e) {
+
+            accountsFeedback.setText(ERROR_FXML_CONFIG);
+            mainButton.setDisable(true);
+        }
     }
 
     @FXML
-    protected void handleLogin(ActionEvent event) {
-        attemptAuthentication(loginUsername.getText(), loginPassword.getText());
-        loginFeedback.setText(feedback);
+    private void handleLogin(ActionEvent event) {
+        String serverFeedback = server.postAuthenticationRequest(username.getText(),
+                                                                 password.getText());
+        accountsFeedback.setText(serverFeedback);
+        String error = server.getErrorResponse();
+        String response = server.getServerResponse();
+
+        if (serverFeedback.equals(CloudyLauncherServerInterface.ERROR_CONNECTION)) {
+        } else if (!error.isEmpty()) {
+            accountsFeedback.setText(parser.parseErrorResponse(error));
+        } else {
+            token = parser.parseToken(response);
+            setGameDisplayPage();
+        }
     }
 
     @FXML
-    protected void handleGame(ActionEvent event) {
+    private void handleSignup(ActionEvent event) {
+        String serverFeedback = server.postSignupRequest(username.getText(),
+                                                         password.getText(),
+                                                         email.getText());
 
-        Button gameButton = (Button) event.getTarget();
-        if (gameButton.getText().equals("Join Game")) {
+        accountsFeedback.setText(serverFeedback);
+        String error = server.getErrorResponse();
+
+        if (serverFeedback.equals(CloudyLauncherServerInterface.ERROR_CONNECTION)) {
+        } else if (!error.isEmpty()) {
+            accountsFeedback.setText(parser.parseErrorResponse(error));
+        } else {
+        }
+    }
+
+    @FXML
+    private void handleLogout(ActionEvent event) {
+        token = "";
+        clearGamePage();
+        setLoginPage(null);
+    }
+
+    @FXML
+    private void handleGame(ActionEvent event) {
+        if (listOfOwnedGames.contains(selectedGame)) {
             joinGame(selectedGame);
         } else {
-            buyGame(selectedGame);
-            handleRefreshGameList(new ActionEvent());
-        }
-
-        joinFeedback.setText(feedback);
-    }
-
-    @FXML
-    protected void handleLogOut(ActionEvent event) {
-        resetAllValues();
-    }
-
-    @FXML
-    protected void handleDisplayAllGames(ActionEvent event) {
-        gameRoot.getChildren().clear();
-        for (Game game : listOfGames) {
-
-            StackPane gameIcon = createNewGameThumbnail(game);
-            gameRoot.getChildren().add(gameIcon);
+            gameFeedback.setText("obtain game first");
         }
     }
 
     @FXML
-    protected void handleDisplayMyGames(ActionEvent event) {
-        gameRoot.getChildren().clear();
-        for (Game game : listOfGames) {
-            if (listOfOwnedIds.contains(game.getId())) {
-                StackPane gameIcon = createNewGameThumbnail(game);
-                gameRoot.getChildren().add(gameIcon);
+    private void displayAllGames(ActionEvent event) {
+        setGamePanel(GAME_DISPLAY_WELCOME);
+        setPaginationSettings(PAGINATION_ALL_GAMES);
+    }
+
+    @FXML
+    private void displayMyCollection(ActionEvent event) {
+        setGamePanel(GAME_DISPLAY_WELCOME);
+        setPaginationSettings(PAGINATION_MY_GAMES);
+    }
+
+    @FXML
+    private void refreshGameList(ActionEvent event) {
+        clearGamePage();
+        setGameDisplayPage();
+    }
+
+    @FXML
+    private void setSignupPage(MouseEvent event) {
+        if (!mainButton.isDisabled()) {
+            try {
+                FXMLLoader loader = new FXMLLoader(getClass().getResource(PATH_ACCOUNTS_SIGNUP));
+                loader.setController(this);
+                mainContent.setCenter(loader.load());
+
+            } catch (IOException e) {
+                accountsFeedback.setText(ERROR_FXML_SIGNUP);
             }
         }
-        if (gameDisplayLayout.getChildren().contains(gameInfoPanel)) {
-            gameDisplayLayout.getChildren().remove(gameInfoPanel);
-            gameStage.setWidth(gameStage.getWidth() - 250);
-        }
     }
 
     @FXML
-    protected void handleRefreshGameList(ActionEvent event) {
-        resetGameListValues();
-        initialiseGameDisplayPanel();
-    }
-
-    @FXML
-    private void handleDisplayGameInfo(MouseEvent event) {
-
-        Node target = (Node) event.getTarget();
-        String tid = target.getId();
-
-        if (target instanceof ImageView) {
-            if (!gameDisplayLayout.getChildren().contains(gameInfoPanel)) {
-                gameDisplayLayout.getChildren().add(gameInfoPanel);
-                gameStage.setWidth(gameStage.getWidth() + infoWidth);
-            }
-
-            ImageView selectedIcon = (ImageView) event.getTarget();
-            selectedGame = (Game) selectedIcon.getUserData();
-
-            String baseGameInfo = "Publisher: %s\nMaximum number of players: %s\nAvailability: %s";
-
-            gameName.setText(selectedGame.getName());
-            boolean isGameOwned = listOfOwnedIds.contains(selectedGame.getId());
-
-            String availability;
-            if (isGameOwned) {
-                availability = "Owned";
-                gameButton.setText("Join Game");
-            } else {
-                availability = "Not Owned";
-                gameButton.setText("Buy Game");
-            }
-            gameInfo.setText(String.format(baseGameInfo,
-                                           selectedGame.getPublisher(),
-                                           selectedGame.getLimit(), availability));
-
-        } else if (((target instanceof Region) && !(target instanceof VBox))
-                   || (tid.equals("tilePaneBase")) || (tid.equals("gameRoot"))) {
-            if (gameDisplayLayout.getChildren().contains(gameInfoPanel)) {
-                gameDisplayLayout.getChildren().remove(gameInfoPanel);
-                gameStage.setWidth(gameStage.getWidth() - infoWidth);
-            }
-
-        } else {
-        }
-    }
-
-    private void buyGame(Game gameToBuy) {
+    private void setLoginPage(MouseEvent event) {
         try {
-            URL url = new URL(baseurl + "/game-ownership/");
-            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-
-            String auth = "Token " + token;
-
-            connection.setRequestMethod("POST");
-            connection.setRequestProperty("Authorization", auth);
-            connection.setRequestProperty("Accept", "application/json");
-            String queryData = String.format("game=%s&user=%s",
-                                             gameToBuy.getId(),
-                                             loginUsername.getText());
-
-            connection.setDoOutput(true);
-            DataOutputStream writer = new DataOutputStream(connection.getOutputStream());
-
-            writer.writeBytes(queryData);
-            writer.flush();
-            writer.close();
-
-            if (connection.getResponseCode() == HttpURLConnection.HTTP_FORBIDDEN) {
-                setFeedback("Invalid authorization token");
-
-            } else if (connection.getResponseCode() == HttpURLConnection.HTTP_CREATED) {
-                setFeedback("Successfully requested game");
-
-            } else {
-                // if system is working properly, should not reach here
-                setFeedback(connection.getHeaderField(0));
-            }
+            FXMLLoader loader = new FXMLLoader(getClass().getResource(PATH_ACCOUNTS_LOGIN));
+            loader.setController(this);
+            mainContent.setCenter(loader.load());
 
         } catch (IOException e) {
-            setFeedback("Check connection to server.");
+            accountsFeedback.setText(ERROR_FXML_LOGIN);
+            mainButton.setDisable(true);
         }
     }
 
-    private void joinGame(Game gameToJoin) {
-
+    private void setGameDisplayPage() {
         try {
-            String controllerId = getControllerId(gameToJoin);
-            String streamingPort;
+            FXMLLoader loader = new FXMLLoader(getClass().getResource(PATH_GAME_DISPLAY_BASE));
+            loader.setController(this);
+            mainContent.setCenter(loader.load());
 
-            if (controllerId == "-1") {
-                joinFeedback.setText(feedback);
-            } else {
-                streamingPort = getStreamingPort(gameToJoin);
-                if (streamingPort == "-1") {
-                    joinFeedback.setText(feedback);
+            setGamePanel(GAME_DISPLAY_WELCOME);
 
-                } else {
+            initialiseGameList();
+            initialiseOwnedGameList();
+            displayAllGames(null);
 
-                    String runThinClientCommand = String.format("python thin_client/main.py %s %s %s",
-                                                                gameToJoin.getAddress(),
-                                                                streamingPort, controllerId);
-                    Runtime.getRuntime().exec(runThinClientCommand);
+        } catch (IOException e) {
+            gameFeedback.setText(ERROR_FXML_GAME_DISPLAY);
+            gameButton.setDisable(true);
+        }
+    }
+
+    @FXML
+    private void setSettingsPage(ActionEvent event) {
+        System.out.println("settings");
+    }
+
+    private void initialiseLauncher(Stage stage) {
+        try {
+            FXMLLoader loader = new FXMLLoader(getClass().getResource(PATH_LAUNCHER_BASE));
+            loader.setController(this);
+            stage = loader.load();
+
+            loader = new FXMLLoader(getClass().getResource(PATH_ACCOUNTS_LOGIN));
+            loader.setController(this);
+            mainContent.setCenter(loader.load());
+
+            setupConfig();
+
+            stage.show();
+        } catch (IOException e) {
+            accountsFeedback.setText(ERROR_FXML_LOGIN);
+            mainButton.setDisable(true);
+        }
+    }
+
+    private void initialiseGameList() {
+        String serverFeedback = server.postAllGamesQuery(token);
+        gameFeedback.setText(serverFeedback);
+        String error = server.getErrorResponse();
+        String response = server.getServerResponse();
+
+        if (serverFeedback.equals(CloudyLauncherServerInterface.ERROR_CONNECTION)) {
+        } else if (!error.isEmpty()) {
+            gameFeedback.setText(parser.parseErrorResponse(error));
+        } else {
+            initialiseGameListFromList(parser.parseGameList(response));
+        }
+    }
+
+    private void initialiseOwnedGameList() {
+        String serverFeedback = server.postOwnedGamesQuery(username.getText(),
+                                                           token);
+        gameFeedback.setText(serverFeedback);
+        String error = server.getErrorResponse();
+        String response = server.getServerResponse();
+        List<String> listOfOwnedIds;
+
+        if (serverFeedback.equals(CloudyLauncherServerInterface.ERROR_CONNECTION)) {
+        } else if (!error.isEmpty()) {
+            gameFeedback.setText(parser.parseErrorResponse(error));
+        } else {
+            listOfOwnedIds = parser.parseOwnedIdList(response);
+
+            for (Game game : listOfGames) {
+                if (listOfOwnedIds.contains(game.getId())) {
+                    listOfOwnedGames.add(game);
                 }
             }
-        } catch (IOException e) {
-            setFeedback("Error joining game");
-            joinFeedback.setText(feedback);
         }
+    }
+
+    private void initialiseGameListFromList(List<Map<CloudyLauncherJsonParser.GameInformation, String>> gameList) {
+        if (gameList.isEmpty()) {
+            System.out.println("game list error");
+
+        } else {
+            for (Map<CloudyLauncherJsonParser.GameInformation, String> game : gameList) {
+                Game newGame = new Game(game.get(CloudyLauncherJsonParser.GameInformation.ID),
+                                        game.get(CloudyLauncherJsonParser.GameInformation.NAME),
+                                        game.get(CloudyLauncherJsonParser.GameInformation.PUBLISHER),
+                                        Integer.parseInt(game.get(CloudyLauncherJsonParser.GameInformation.LIMIT)),
+                                        game.get(CloudyLauncherJsonParser.GameInformation.ADDRESS),
+                                        game.get(CloudyLauncherJsonParser.GameInformation.THUMBNAIL));
+                listOfGames.add(newGame);
+            }
+        }
+    }
+
+    private void setGamePanel(Integer gameListType) {
+        try {
+            FXMLLoader loader;
+
+            if (gameListType.equals(GAME_DISPLAY_WELCOME)) {
+                loader = new FXMLLoader(getClass().getResource(PATH_GAME_DISPLAY_INITIAL));
+                gameButton.setDisable(true);
+            } else {
+                loader = new FXMLLoader(getClass().getResource(PATH_GAME_DISPLAY_GAME));
+                gameButton.setDisable(false);
+            }
+
+            loader.setController(this);
+            VBox gamePanelContent = loader.load();
+
+            gameInfoPanel.getChildren().clear();
+            gameInfoPanel.getChildren().addAll(gamePanelContent.getChildren());
+
+        } catch (IOException e) {
+            gameFeedback.setText(ERROR_FXML_GAME_DISPLAY);
+            gameButton.setDisable(true);
+        }
+    }
+
+    private void setGameInformation(MouseEvent event) {
+
+        setGamePanel(GAME_DISPLAY_GAME_INFO);
+
+        ImageView selectedIcon = (ImageView) event.getTarget();
+        selectedGame = (Game) selectedIcon.getUserData();
+
+        gameTitle.setText(selectedGame.getName());
+        gameImage.setImage(selectedIcon.getImage());
+        gameTextInformation.setText(String.format(GAME_INFORMATION,
+                                                  selectedGame.getId(),
+                                                  selectedGame.getPublisher(),
+                                                  selectedGame.getLimit()));
+
+    }
+
+    private void setPaginationSettings(Integer gameListType) {
+
+        int numPages;
+        List<Game> gameList;
+
+        if (gameListType.equals(PAGINATION_ALL_GAMES)) {
+            numPages = listOfGames.size() / TILES_PER_PAGE + 1;
+            gameList = listOfGames;
+        } else {
+            numPages = listOfOwnedGames.size() / TILES_PER_PAGE + 1;
+            gameList = listOfOwnedGames;
+        }
+
+        TilePane gameRoot = new TilePane();
+        gameRoot.getStyleClass().add("game-root");
+
+        pagination.setPageCount(numPages);
+        pagination.setPageFactory(new Callback<Integer, Node>() {
+
+            @Override
+            public Node call(Integer pageIndex) {
+                int startIndex = pageIndex * TILES_PER_PAGE;
+                for (int i = startIndex; i < startIndex + TILES_PER_PAGE; i++) {
+                    if (i < gameList.size()) {
+                        StackPane gameIcon = createNewGameThumbnail(gameList.get(i));
+                        gameRoot.getChildren().add(gameIcon);
+                    } else {
+                        break;
+                    }
+                }
+                return gameRoot;
+            }
+        });
     }
 
     private StackPane createNewGameThumbnail(Game gameInfo) {
@@ -247,14 +369,15 @@ public class CloudyLauncher extends Application {
         gameIcon.setOnMouseClicked(new EventHandler<MouseEvent>() {
             @Override
             public void handle(MouseEvent event) {
-                handleDisplayGameInfo(event);
+                setGameInformation(event);
             }
         });
 
-        ImageView icon = new ImageView("images/orangeribbon.png");
+        ImageView icon = new ImageView(URL_OWNED_BADGE);
         StackPane img = new StackPane();
         img.getStyleClass().add("game-icon");
-        if (listOfOwnedIds.contains(gameInfo.getId())) {
+
+        if (listOfOwnedGames.contains(gameInfo)) {
             img.getChildren().addAll(gameIcon, icon);
         } else {
             img.getChildren().addAll(gameIcon);
@@ -263,451 +386,50 @@ public class CloudyLauncher extends Application {
         return img;
     }
 
-    private void setToken(String newToken) {
-        token = newToken;
-    }
-
-    private void setFeedback(String newFeedback) {
-        feedback = newFeedback;
-    }
-
-    private void attemptAuthentication(String username, String password) {
-
-        try {
-            URL url = new URL(baseurl + "/api-token-auth/");
-            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-
-            connection.setRequestMethod("POST");
-            connection.setRequestProperty("Accept", "application/json");
-            String queryData = String.format("username=%s&password=%s", username, password);
-
-            connection.setDoOutput(true);
-            DataOutputStream writer = new DataOutputStream(connection.getOutputStream());
-
-            writer.writeBytes(queryData);
-            writer.flush();
-            writer.close();
-
-            if (connection.getResponseCode() == HttpURLConnection.HTTP_BAD_REQUEST) {
-                setErrorMessageFromConnection(connection);
-
-            } else if (connection.getResponseCode() == HttpURLConnection.HTTP_OK){
-
-                BufferedReader reader = new BufferedReader(new InputStreamReader(connection.getInputStream()));
-                String response = reader.readLine();
-                reader.close();
-
-                JSONObject tokenObj = new JSONObject(response);
-                String responseToken = tokenObj.getString("token");
-                setToken(responseToken);
-
-                setFeedback("User recognised.");
-
-                initialiseGameDisplayPanel();
-
-            } else {
-                setFeedback(connection.getHeaderField(0));
-            }
-
-        } catch (IOException e) {
-            setFeedback("Check connection to server.");
-        }
-    }
-
-    private void attemptUserRegistration(String username, String password,
-                                         String email) {
-
-        try {
-            URL url = new URL(baseurl + "/users/");
-            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-
-            connection.setRequestMethod("POST");
-            connection.setRequestProperty("Accept", "application/json");
-            String queryData = String.format("username=%s&password=%s&email=%s",
-                                             username, password, email);
-
-            connection.setDoOutput(true);
-            DataOutputStream writer = new DataOutputStream(connection.getOutputStream());
-
-            writer.writeBytes(queryData);
-            writer.flush();
-            writer.close();
-
-            if (connection.getResponseCode() == HttpURLConnection.HTTP_BAD_REQUEST) {
-                setErrorMessageFromConnection(connection);
-
-            } else if (connection.getResponseCode() == HttpURLConnection.HTTP_CREATED){
-                setFeedback("User successfully registered.");
-
-            } else {
-                setFeedback(connection.getHeaderField(0));
-            }
-
-        } catch (IOException e) {
-            setFeedback("Check connection to server.");
-        }
-    }
-
-    private void setErrorMessageFromConnection(HttpURLConnection connection) {
-        String errorMessage = "";
-
-        BufferedReader errorReader = new BufferedReader(new InputStreamReader(connection.getErrorStream()));
-        JSONObject errorResponse;
-        try {
-            errorResponse = new JSONObject(errorReader.readLine());
-
-            if (errorResponse.has("username")) {
-                errorMessage = errorMessage
-                               + "\nUsername: "
-                               + errorResponse.getJSONArray("username")
-                                              .getString(0);
-            }
-
-            if (errorResponse.has("password")) {
-                errorMessage = errorMessage
-                               + "\nPassword: "
-                               + errorResponse.getJSONArray("password")
-                                              .getString(0);
-            }
-
-            if (errorResponse.has("non_field_errors")) {
-                errorMessage = errorMessage
-                               + "\n"
-                               + errorResponse.getJSONArray("non_field_errors")
-                                              .getString(0);
-            }
-
-            if (errorResponse.has("email")) {
-                errorMessage = errorMessage
-                               + "\nEmail: "
-                               + errorResponse.getJSONArray("email")
-                                              .getString(0);
-            }
-
-            if (errorResponse.has("first_name")) {
-                errorMessage = errorMessage
-                               + "\nFirst Name: "
-                               + errorResponse.getJSONArray("first_name")
-                                              .getString(0);
-            }
-
-            if (errorResponse.has("last_name")) {
-                errorMessage = errorMessage
-                               + "\nLast Name: "
-                               + errorResponse.getJSONArray("last_name")
-                                              .getString(0);
-            }
-
-            if (errorResponse.has("detail")) {
-                errorMessage = errorResponse.getJSONArray("username")
-                                            .getString(0);
-            }
-
-            if (errorResponse.has("game")) {
-                errorMessage = errorMessage
-                               + "\nGame: "
-                               + errorResponse.getJSONArray("game")
-                                              .getString(0);
-            }
-
-            if (errorResponse.has("controller")) {
-                errorMessage = errorMessage
-                               + "\nController: "
-                               + errorResponse.getJSONArray("controller")
-                                              .getString(0);
-            }
-
-            if (errorResponse.has("player")) {
-                errorMessage = errorMessage
-                               + "\nPlayer: "
-                               + errorResponse.getJSONArray("player")
-                                              .getString(0);
-            }
-
-            setFeedback(errorMessage);
-            errorReader.close();
-
-        } catch (JSONException e) {
-            setFeedback("Error parsing the response from server");
-
-        } catch (IOException e) {
-            setFeedback("Error accessing the error stream");
-        }
-    }
-
-    private Game getGameFromJson(JSONObject gameObj) {
-
-        try {
-            String gId = Integer.toString(gameObj.getInt("id"));
-            String gName = gameObj.getString("name");
-            String gPublisher = gameObj.getString("publisher");
-            String gLimit = Integer.toString(gameObj.getInt("max_limit"));
-            String gAddress = gameObj.getString("address");
-            String gThumbnail = gameObj.getString("thumbnail");
-
-            Game newGame = new Game(gId, gName, gPublisher,
-                                    Integer.parseInt(gLimit), gAddress, gThumbnail);
-            return newGame;
-
-        } catch (JSONException e) {
-            gameInfo.setText("Error in parsing game information");
-
-        }
-        return null;
-    }
-
-    private String getStreamingPort(Game gameToJoin) {
-        try {
-            URL url = new URL(baseurl + String.format("/game-session/?game=%s&user=%s",
-                                                      gameToJoin.getId(),
-                                                      loginUsername.getText()));
-            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-
-            String auth = "Token " + token;
-
-            connection.setRequestMethod("GET");
-            connection.setRequestProperty("Authorization", auth);
-            connection.setRequestProperty("Accept", "application/json");
-
-            if (connection.getResponseCode() == HttpURLConnection.HTTP_OK) {
-
-                BufferedReader reader = new BufferedReader(new InputStreamReader(connection.getInputStream()));
-                String response = reader.readLine();
-                reader.close();
-
-                JSONObject gameSession = new JSONArray(response).getJSONObject(0);
-                try {
-                    String portNum = Integer.toString(gameSession.getInt("streaming_port"));
-                    return portNum;
-                } catch (JSONException e) {
-                    setFeedback("Error parsing port number from server");
-                }
-
-            } else {
-                // if system is working properly, should not reach here
-                setFeedback(connection.getHeaderField(0));
-            }
-
-        } catch (IOException e) {
-            setFeedback("Check connection to server.");
-        }
-
-        return "-1";
-    }
-    private String getControllerId(Game gameToJoin) {
-        try {
-            URL url = new URL(baseurl + "/game-session/");
-            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-
-            String auth = "Token " + token;
-
-            connection.setRequestMethod("POST");
-            connection.setRequestProperty("Authorization", auth);
-            connection.setRequestProperty("Accept", "application/json");
-            String queryData = String.format("game=%s&user=%s",
-                                             gameToJoin.getId(),
-                                             loginUsername.getText());
-
-            connection.setDoOutput(true);
-            DataOutputStream writer = new DataOutputStream(connection.getOutputStream());
-
-            writer.writeBytes(queryData);
-            writer.flush();
-            writer.close();
-
-            if (connection.getResponseCode() == HttpURLConnection.HTTP_BAD_REQUEST) {
-                System.out.println("bad request: possibly wrong data format passed,"
-                        + "possibly session already exists");
-
-            } else if (connection.getResponseCode() == HttpURLConnection.HTTP_CREATED) {
-
-                BufferedReader reader = new BufferedReader(new InputStreamReader(connection.getInputStream()));
-                String response = reader.readLine();
-                reader.close();
-
-                JSONObject gameSession = new JSONObject(response);
-
-                try {
-                    String controllerId = Integer.toString(gameSession.getInt("controller"));
-                    return controllerId;
-                } catch (JSONException e) {
-                    setFeedback("Error parsing controller id from server");
-                }
-
-            } else {
-                // if system is working properly, should not reach here
-                setFeedback(connection.getHeaderField(0));
-            }
-
-        } catch (IOException e) {
-            setFeedback("Check connection to server.");
-        }
-
-        return "-1";
-    }
-
-    private void initialiseGameList() {
-        try {
-            URL url = new URL(baseurl + "/games/");
-            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-
-            String auth = "Token " + token;
-
-            connection.setRequestMethod("GET");
-            connection.setRequestProperty("Authorization", auth);
-            connection.setRequestProperty("Accept", "application/json");
-
-            BufferedReader reader = new BufferedReader(new InputStreamReader(connection.getInputStream()));
-            String response = reader.readLine();
-            reader.close();
-
-            if (connection.getResponseCode() == HttpURLConnection.HTTP_FORBIDDEN) {
-                // if login (and token retrieval) is successful, should never reach here.
-
-                setErrorMessageFromConnection(connection);
-
-            } else if (connection.getResponseCode() == HttpURLConnection.HTTP_OK) {
-                setFeedback("token recognised.");
-
-                JSONArray gameListString = new JSONArray(response);
-
-                for (int i = 0; i < gameListString.length(); i++) {
-                    JSONObject gameJSON = gameListString.getJSONObject(i);
-                    Game game = getGameFromJson(gameJSON);
-                    listOfGames.add(game);
-                }
-
-            } else {
-                setFeedback(connection.getHeaderField(0));
-            }
-
-        } catch (IOException e) {
-            setFeedback("Check connection to server.");
-        }
-    }
-
-    private void getOwnedGamesList() {
-        try {
-            URL url = new URL(baseurl + "/game-ownership/?user=" + loginUsername.getText());
-            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-
-            String auth = "Token " + token;
-
-            connection.setRequestMethod("GET");
-            connection.setRequestProperty("Authorization", auth);
-            connection.setRequestProperty("Accept", "application/json");
-
-            BufferedReader reader = new BufferedReader(new InputStreamReader(connection.getInputStream()));
-            String response = reader.readLine();
-            reader.close();
-
-            if (connection.getResponseCode() == HttpURLConnection.HTTP_FORBIDDEN) {
-                System.out.println("invalid token");
-                setErrorMessageFromConnection(connection);
-
-            } else if (connection.getResponseCode() == HttpURLConnection.HTTP_OK) {
-                setFeedback("token recognised.");
-
-                JSONArray gameOwnedString = new JSONArray(response);
-
-                for (int i = 0; i < gameOwnedString.length(); i++) {
-                    JSONObject gameJSON = gameOwnedString.getJSONObject(i);
-                    try {
-                        String gId = Integer.toString(gameJSON.getInt("game"));
-                        listOfOwnedIds.add(gId);
-
-                    } catch (JSONException e) {
-                        gameInfo.setText("Error in parsing owned games");
-                    }
-                }
-
-            } else {
-                setFeedback(connection.getHeaderField(0));
-            }
-
-        } catch (IOException e) {
-            setFeedback("Check connection to server.");
-        }
-    }
-
-    private void initialiseGameDisplayPanel() {
-
-        initialiseGameList();
-        getOwnedGamesList();
-        handleDisplayAllGames(new ActionEvent());
-
-        try {
-            Scene scene = new Scene(gameDisplayLayout, 500, 300);
-            gameStage.setScene(scene);
-
-        } catch (IllegalArgumentException e) {
-            // gameStage already initialised
-        }
-
-        if (gameDisplayLayout.getChildren().contains(gameInfoPanel)) {
-            gameDisplayLayout.getChildren().remove(gameInfoPanel);
-        }
-
-        userStage.hide();
-        gameStage.show();
-        gameStage.setMinWidth(500);
-        gameStage.setMinHeight(300);
-        gameStage.sizeToScene();
-    }
-
-    private void resetAllValues() {
-
-        token = "";
-        feedback = "";
-
-        signupEmail.clear();
-        signupUsername.clear();
-        signupPassword.clear();
-        signupFeedback.setText("");
-
-        loginUsername.clear();
-        loginPassword.clear();
-        loginFeedback.setText("");
-
-        resetGameListValues();
-
-        userStage.show();
-        gameStage.hide();
-    }
-
-    private void resetGameListValues() {
+    private void clearGamePage() {
         listOfGames.clear();
-        listOfOwnedIds.clear();
-        gameInfo.setText("");
-        gameRoot.getChildren().clear();
-        joinFeedback.setText("");
+        listOfOwnedGames.clear();
     }
 
-    private void initialise() {
-        try {
-            rootLayout = (VBox) FXMLLoader.load(getClass().getResource("CloudyLauncher.fxml"));
-            ObservableList<Node> rChildren = rootLayout.getChildren();
+    private void joinGame(Game gameToJoin) {
 
-            rChildren.remove(1, rChildren.size());
+        String serverFeedback = server.postGameSessionRequest(gameToJoin,
+                                                              username.getText(),
+                                                              token);
 
-        } catch (IOException e) {
-            System.out.println("Error in loading fxml file");
+        gameFeedback.setText(serverFeedback);
+        String error = server.getErrorResponse();
+        String response = server.getServerResponse();
+
+        if (serverFeedback.equals(CloudyLauncherServerInterface.ERROR_CONNECTION)) {
+        } else if (!error.isEmpty()) {
+            gameFeedback.setText(parser.parseErrorResponse(error));
+        } else {
+            Map<CloudyLauncherJsonParser.GameSession, String> gameSession = parser.parseGameSession(response);
+            String controllerId = gameSession.get(CloudyLauncherJsonParser.GameSession.CONTROLLER);
+            String port = gameSession.get(CloudyLauncherJsonParser.GameSession.PORT);
+            String sessionId = gameSession.get(CloudyLauncherJsonParser.GameSession.ID);
+
+            try {
+                Runtime.getRuntime()
+                       .exec(String.format(COMMAND_RUN_THINCLIENT,
+                                           gameToJoin.getAddress(), port,
+                                           controllerId, sessionId));
+            } catch (IOException e) {
+                gameFeedback.setText(ERROR_GAME_JOIN);
+            }
         }
     }
 
     @Override
     public void start(Stage primaryStage) throws Exception {
 
-        initialise();
-        Scene scene = new Scene(rootLayout, 320, 300);
-
-        userStage.setResizable(false);
-        userStage.setTitle("CloudyLauncher");
-        userStage.setScene(scene);
-        userStage.show();
+        CloudyLauncher cl = new CloudyLauncher();
+        cl.initialiseLauncher(primaryStage);
     }
 
     public static void main(String[] args) {
         launch(args);
     }
+
 }
