--- style.css
@@ -1,114 +1,0 @@
-.root {
-    -fx-background-color: #023350;
-}
-
-.user-panel {
-    -fx-alignment: top-center;
-    -fx-min-height: 400;
-    -fx-padding: 20;
-    -fx-spacing: 20;
-}
-
-.grid-pane {
-    -fx-alignment: center;
-    -fx-hgap: 20;
-    -fx-vgap: 10;
-    -fx-text-fill: #FDCCAF;
-}
-
-.buttons-display {
-    -fx-spacing: 10;
-}
-
-.button {
-    -fx-background-color: #0D4E75;
-    -fx-text-fill: #D98F63;
-    -fx-font-family: "trajan pro";
-}
-
-.button:hover {
-    -fx-background-color: #6DA9CC;
-    -fx-text-fill: #925633;
-}
-
-.tab-pane {
-    -fx-tab-min-width: 120;
-}
-
-.tab-header-background{
-    -fx-background-color: #0D4E75;
-}
-
-.headers-region{
-    -fx-color: #D98F63;
-    -fx-font-weight: bold;
-    -fx-font-family: "trajan pro";
-}
-
-.tab:selected {
-    -fx-background-color: #6DA9CC;
-}
-
-.tab:hover{
-    -fx-background-color: #6DA9CC;
-}
-
-.tab-label{
-    -fx-text-fill: #925633;
-}
-
-.label {
-    -fx-text-fill: #FDCCAF;
-    -fx-font-size: 15;
-}
-
-.feedback {
-    -fx-fill: #F28A8A;
-    -fx-font-weight: bold;
-    -fx-height: 100;
-    -fx-min-height: 100;
-}
-
-.text-field {
-    -fx-background-color: #6DA9CC;
-    -fx-highlight-text-fill: #F2B18A;
-    -fx-prompt-text-fill: #925633;
-}
-
-.refresh-button {
-    -fx-background-image: url("images/refresh.png");
-    -fx-min-width: 25;
-}
-
-.game-tile-layout {
-    -fx-alignment: top-right;
-    -fx-padding: 10;
-    -fx-spacing: 20;
-}
-
-.game-root {
-    -fx-alignment: center;
-    -fx-hgap: 10;
-    -fx-vgap: 10;
-}
-
-.game-icon {
-    -fx-alignment: top-right;
-}
-
-.game-info-name {
-    -fx-fill: #FDCCAF;
-    -fx-font-family: "Arial Narrow";
-    -fx-font-weight: bold;
-    -fx-font-size: 25;
-    -fx-underline: true;
-}
-
-.game-info {
-    -fx-alignment: center;
-    -fx-spacing: 20;
-    -fx-min-width: 250;
-    -fx-border-width: 0 0 0 5;
-    -fx-border-color: radial-gradient(center 50% 50%, radius 50%,  #023350, #925633, #023350);
-    -fx-border-style: solid;
-}
